.pyrunner/requirements.txt
pillow
pandastable